longtask-timing/longtask-executescript.html

<!DOCTYPE HTML>
<meta charset=utf-8>
<title>LongTask Timing: validate the long task attributes for execute-script</title>
<body>

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<h1>Long Task Attributes for Execute-Script</h1>
<div id="log"></div>
<script>
  async_test(function (t) {
    const observer = new PerformanceObserver(
      t.step_func(function (entryList) {
        const entries = entryList.getEntries();
        assert_equals(entries.length, 1,
            'Exactly one entry is expected.');
        const longtask = entries[0];
        assert_equals(longtask.entryType, 'longtask');
        assert_equals(longtask.name, 'self');
        assert_greater_than(longtask.duration, 50);
        assert_greater_than_equal(longtask.startTime, 0);
        const currentTime = performance.now();
        assert_less_than_equal(longtask.startTime, currentTime);

        var exist = false;
        for (const attribution of longtask.attribution) {
          if (attribution.name !== 'script-run' || attribution.duration < 51)
            continue;
          exist = true;
          assert_greater_than(attribution.startTime, 0);
          assert_equals(attribution.entryType, 'taskattribution');
          assert_equals(attribution.containerId, '');
          assert_equals(attribution.containerName, '');
          assert_equals(attribution.containerSrc, '');
        }
        assert_equals(exist, true);
        observer.disconnect();
        t.done();
      })
    );
    observer.observe({entryTypes: ['longtask']});

    /* Generate a slow task */
    const begin = window.performance.now();
    while (window.performance.now() < begin + 51);

}, 'Performance longtask entries with execute-script attribute are observable.');
</script>

</body>
